👌 IMPROVE: Move common components in aurelius package to common directory
