+++ packages/aurelius/src/components/index.tsx
@@ -1,2 +1,2 @@
 export { default as Writer } from './writer'
-export { default as Settings } from './settings'
+export * from './common'
